refactor(brutus): extract input loading to input.go
Move loadPasswords and loadKey functions.
Part of main.go refactoring (6/9).

Related: C1/C2 (split 1340-line main.go)
